Refactor naming of Account model's namespace
Referring to Account model with previous naming was confusing.
