+++ GentlemensClub-Backend/Views/Account/Login.cshtml
@@ -1,4 +1,4 @@
-﻿@model GentlemensClub.Models.Account.LoginCredential
+﻿@model GentlemensClub.Models.Authentication.LoginCredential
 @{
 }
 <div class="card container p-4 bg-black" style="width: 24rem">
