+++ GentlemensClub-Backend/Views/Account/Registration.cshtml
@@ -1,4 +1,4 @@
-﻿@model GentlemensClub.Models.Account.RegistrationData
+﻿@model GentlemensClub.Models.Authentication.RegistrationData
 @{
 }
 <div class="card container p-4 bg-black" style="width: 24rem">
